+++ Classes/Geocoding/Google.php
@@ -124,8 +124,6 @@ class Tx_Oelib_Geocoding_Google implements \Tx_Oelib_Interface_GeocodingLookup
      * @param \Tx_Oelib_Interface_Geo $geoObject
      *
      * @return void
-     *
-     * @throws \RuntimeException
      */
     public function lookUp(\Tx_Oelib_Interface_Geo $geoObject)
     {
@@ -143,13 +141,14 @@ class Tx_Oelib_Geocoding_Google implements \Tx_Oelib_Interface_GeocodingLookup
         $attempts = 0;
         do {
             $this->throttle($throttleTime);
-            $lookupError = false;
 
             $retry = false;
             $response = $this->sendRequest($address);
 
-            $httpError = $response === false;
-            if (!$httpError) {
+            $lookupError = $response === false;
+            if ($lookupError) {
+                $status = 'network problem';
+            } else {
                 $resultParts = json_decode($response, true);
                 $status = $resultParts['status'];
                 $lookupError = $status !== self::STATUS_OK;
@@ -162,17 +161,13 @@ class Tx_Oelib_Geocoding_Google implements \Tx_Oelib_Interface_GeocodingLookup
                 }
             }
 
-            if ($httpError || $lookupError) {
+            if ($lookupError) {
                 $attempts++;
                 if ($attempts < static::MAXIMUM_ATTEMPTS) {
                     $retry = true;
                 }
             }
         } while ($retry);
-
-        if ($httpError) {
-            throw new \RuntimeException('There was an error connecting to the Google geocoding server.', 1331488446);
-        }
 
         if (!$lookupError) {
             $coordinates = $resultParts['results'][0]['geometry']['location'];
@@ -183,7 +178,7 @@ class Tx_Oelib_Geocoding_Google implements \Tx_Oelib_Interface_GeocodingLookup
                 ]
             );
         } else {
-            $geoObject->setGeoError();
+            $geoObject->setGeoError($status);
         }
     }