+++ Tests/Unit/Geocoding/GoogleTest.php
@@ -149,7 +149,7 @@ class GoogleTest extends \Tx_Phpunit_TestCase
      *
      * @dataProvider noResultsStatusDataProvider
      */
-    public function lookUpForAFullGermanAddressWithNoCoordinatesFoundSetsGeoProblem($status)
+    public function lookUpForAFullGermanAddressWithNoCoordinatesFoundSetsGeoProblemAndLogsError($status)
     {
         $jsonResult = '{ "status": "' . $status . '" }';
 
@@ -169,14 +169,13 @@ class GoogleTest extends \Tx_Phpunit_TestCase
         $subject->lookUp($geo);
 
         self::assertTrue($geo->hasGeoError());
+        self::assertSame($status, $geo->getGeoErrorReason());
     }
 
     /**
      * @test
-     *
-     * @expectedException \RuntimeException
      */
-    public function lookUpForAFullGermanAddressWithNetworkErrorThrowsException()
+    public function lookUpForAFullGermanAddressWithNetworkErrorSetsGeoProblemAndLogsError()
     {
         $geo = new \Tx_Oelib_Tests_Unit_Fixtures_TestingGeo();
         $geo->setGeoAddress('Am Hof 1, 53113 Zentrum, Bonn, DE');
@@ -192,6 +191,9 @@ class GoogleTest extends \Tx_Phpunit_TestCase
         $subject->expects(self::any())->method('sendRequest')->will(self::returnValue(false));
 
         $subject->lookUp($geo);
+
+        self::assertTrue($geo->hasGeoError());
+        self::assertSame('network problem', $geo->getGeoErrorReason());
     }
 
     /**